requirements.txt
click==8.1.7
cmake==3.29.3
dlib==19.9.0
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.14.0
fsspec==2024.5.0
greenlet==3.0.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==2.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
pillow==10.3.0
pip==24.0
psycopg2==2.9.9
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
setuptools==69.5.1
SQLAlchemy==2.0.30
sympy==1.12
torch==2.3.0
torchvision==0.18.0
triton==2.3.0
typing_extensions==4.11.0
wheel==0.43.0
